refactor(audio): extract ExponentialEnvelope into envelope.h
Move ExponentialEnvelope out of realtime_runner.h into its own header
(core/include/magentart/envelope.h) so it can be included and tested
independently without pulling in the full MLX/TFLite machinery.

- envelope.h: defines ExponentialEnvelope in magentart::core namespace
- realtime_runner.h: includes envelope.h, struct definition removed
- envelope_reset_test.cpp: includes envelope.h directly instead of
  mirroring the struct -- test now covers the real production definition
- CMakeLists.txt: add include path for envelope_reset_test target
